Split MessagesHandler to MessagesHandler and NetworkMessageType.
